requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
openai==1.50.0
python-dotenv==1.0.1
pydantic==2.9.0
